wip: optional types

+++ src/ast/ast_simulate.cpp
@@ -3174,7 +3174,7 @@ namespace das
             }
         }
         context.totalInitFunctions = (uint32_t) allInitFunctions.size();
-        context.initFunctions = (SimFunction **) context.code->allocate(allInitFunctions.size()*sizeof(SimFunction *));
+        context.initFunctions = (SimFunction **) context.code->allocate(uint32_t(allInitFunctions.size()*sizeof(SimFunction *)));
         memcpy ( context.initFunctions, allInitFunctions.data(), allInitFunctions.size()*sizeof(SimFunction *) );
         // lockchecking
         context.skipLockChecks = options.getBoolOption("skip_lock_checks",false);
